fix(*): resolve type errors, fix test expectations, and remove find-up dependency
Add branded type casts (Int, Positive, FragmentStringJsonc) at assertion sites across
13 test files to satisfy exactOptionalPropertyTypes with bun:test matchers.

Fix PlainJsonBase.json type from UnknownRecord to JsonValue (genuine type bug in JSONC types).

Correct 4 startsWithComment test expectations: whitespace trimming, end-of-input line
comments, and block comment */ termination per JSONC spec.

Fix regex trim test: replace /Test+/gi with /Test/gi to avoid greedy t+ crossing word
boundaries with case-insensitive flag. Update TSDoc examples in trim, trimStart, trimEnd.

Fix ensureGlobal tests: correct special characters source expectation, document Bun/JSC
unicode escaping behavior in .source property.

Remove find-up dependency from task-util: replace with import.meta.dirname for package
root resolution. Fix shell expansion in special characters test and add permission
enforcement probe for write-permissions test.

+++ packages/module/es/src/types/t object/t array/t p number finite int/f/t number/exclusiveRange/r s/p n/index.unit.test.ts
@@ -6,6 +6,8 @@ import {
 
 import { types, } from '@monochromatic-dev/module-es';
 
+import type { $ as Int, } from '@_/types/t number/t finite/t int/t/index.ts';
+
 const $ = types.object.array.numberFiniteInt.from.number.exclusiveRange.sync.named.$;
 
 describe($, () => {
@@ -18,30 +20,30 @@ describe($, () => {
   });
 
   test('returns correct range for positive integers', () => {
-    expect($({ startExclusive: 0, endExclusive: 5, },),).toEqual([1, 2, 3, 4,],);
+    expect($({ startExclusive: 0, endExclusive: 5, },),).toEqual([1 as Int, 2 as Int, 3 as Int, 4 as Int,],);
   });
 
   test('returns correct range for negative integers', () => {
-    expect($({ startExclusive: -5, endExclusive: 0, },),).toEqual([-4, -3, -2, -1,],);
+    expect($({ startExclusive: -5, endExclusive: 0, },),).toEqual([-4 as Int, -3 as Int, -2 as Int, -1 as Int,],);
   });
 
   test('returns correct range crossing zero', () => {
-    expect($({ startExclusive: -2, endExclusive: 3, },),).toEqual([-1, 0, 1, 2,],);
+    expect($({ startExclusive: -2, endExclusive: 3, },),).toEqual([-1 as Int, 0 as Int, 1 as Int, 2 as Int,],);
   });
 
   test('handles single element ranges', () => {
-    expect($({ startExclusive: 0, endExclusive: 2, },),).toEqual([1,],);
-    expect($({ startExclusive: -1, endExclusive: 1, },),).toEqual([0,],);
+    expect($({ startExclusive: 0, endExclusive: 2, },),).toEqual([1 as Int,],);
+    expect($({ startExclusive: -1, endExclusive: 1, },),).toEqual([0 as Int,],);
   });
 
   test('handles decimal boundaries', () => {
-    expect($({ startExclusive: 0.1, endExclusive: 4.9, },),).toEqual([1, 2, 3, 4,],);
-    expect($({ startExclusive: -0.9, endExclusive: 0.9, },),).toEqual([0,],);
-    expect($({ startExclusive: 1.5, endExclusive: 2.3, },),).toEqual([2,],);
+    expect($({ startExclusive: 0.1, endExclusive: 4.9, },),).toEqual([1 as Int, 2 as Int, 3 as Int, 4 as Int,],);
+    expect($({ startExclusive: -0.9, endExclusive: 0.9, },),).toEqual([0 as Int,],);
+    expect($({ startExclusive: 1.5, endExclusive: 2.3, },),).toEqual([2 as Int,],);
   });
 
   test('handles edge cases with large numbers', () => {
-    expect($({ startExclusive: 99, endExclusive: 103, },),).toEqual([100, 101, 102,],);
+    expect($({ startExclusive: 99, endExclusive: 103, },),).toEqual([100 as Int, 101 as Int, 102 as Int,],);
   });
 
   test('returns empty array when no integers exist in range', () => {
@@ -49,6 +51,6 @@ describe($, () => {
   });
 
   test('handles decimal boundaries that resolve to a single integer', () => {
-    expect($({ startExclusive: 0.5, endExclusive: 1.5, },),).toEqual([1,],);
+    expect($({ startExclusive: 0.5, endExclusive: 1.5, },),).toEqual([1 as Int,],);
   });
 },);
